ref(rules): Clear Redis buffer after processing

+++ src/sentry/buffer/redis.py
@@ -266,7 +266,7 @@ class RedisBuffer(Buffer):
             decoded_set.append(data_and_timestamp)
         return decoded_set
 
-    def delete_key(self, key: str, min: int, max: int) -> None:
+    def delete_key(self, key: str, min: float, max: float) -> None:
         self._execute_redis_operation(key, RedisOperation.SORTED_SET_DELETE_RANGE, min=min, max=max)
 
     def delete_hash(
